requirements.txt
feedparser==6.0.11
jinja2==3.1.3
requests==2.31.0
google-genai>=0.2.0
